Automatic speech recognition (ASR) pipeline using Deep Neural Networks
completed pipeline will accept raw audio as input and return a predicted transcription of the spoken language:

STEP 1 : is a pre-processing step that converts raw audio to one of two feature representations that are commonly used for ASR.

STEP 2 : is an acoustic model which accepts audio features as input and returns a probability distribution over all potential transcriptions. After learning about the basic types of neural networks that are often used for acoustic modeling, you will engage in your own investigations, to design your own acoustic model!

STEP 3 : in the pipeline takes the output from the acoustic model and returns a predicted transcription.